refactor: 프로필 불러오기 API 를 client.ts 로 리펙토링하고 타입을 분리
